Feature/better-sdk-6 (#3641)
* Changes String method in SDK options types to non-pointer method

* Updates docs to better reflect contained_by, contains, produced_by, produces

* Improves names endpoint in SDK

* Improves chunks endpoint in SDK

* Adds an existing_code block to SDK

* Protects against a concurrent map write

* Numerous improvements to the usesSdk example code

* Fixes (temporarily) parsing of articuated items

* Improves usesSDK

* Cleaning

* Better error message for chifra tokens

* Fixes tests

* Improves error reporting for chifra tokens

* Create a way to clean up input to Unmarshall in SDK

* Cleans up inputs and outputs in articulated items prior to Unmarshal

* Improves usesSdk test cases

* float64s instead of the weird double in auto-code gen

* totalDifficulty (a rawonly field) should be uint256, not wei

* Separates ether from amount for chifra blocks --withdrawal --ether

* Fixes issue #3642

* Fixes issue #3642 related to chifra slurp

* Exports balance with --ether option for chifra state

* Cleans up diff field in chifra tokens

* Documents value for chifra traces

* Updates tests for chifra traces

* Updates CHANGES file

* Fixes tests

* Updates usesSdk test

* Tries to start to capture base types (starting with Gas)

* Starts to get control of base types such as TransactionIndex, BlockNumber, etc.

* Separates out base.Txnum (temporary)

* Captures lognum (temporary)

* Gets control of validatorIndex and withdrawalIndex

* Removes special case MustParseGas

* Moves NOPOS and NOPOSI to base package. Adds NOPOSN = Numeral

* Gains complete control of base.Txnum in all types.

* Grab control of LogIndex

+++ src/apps/chifra/internal/traces/handle_counts.go
@@ -89,7 +89,7 @@ func (opts *TracesOptions) HandleCounts() error {
 				for _, item := range items {
 					counter := types.TraceCount{
 						BlockNumber:      uint64(item.BlockNumber),
-						TransactionIndex: uint64(item.TransactionIndex),
+						TransactionIndex: item.TransactionIndex,
 						TransactionHash:  item.Hash,
 						Timestamp:        item.Timestamp,
 						TracesCnt:        uint64(len(item.Traces)),
